requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pandas==2.1.3
openpyxl==3.1.2
requests==2.31.0
python-decouple==3.8